JAV-524 move some logic from RestProducerInvocation to AbstractRestInvocation
