+++ common/common-rest/src/main/java/io/servicecomb/common/rest/RestProducerInvocation.java
@@ -18,13 +18,6 @@ package io.servicecomb.common.rest;
 
 import java.util.List;
 
-import javax.ws.rs.core.Response.Status;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.servicecomb.common.rest.codec.RestCodec;
-import io.servicecomb.common.rest.definition.RestOperationMeta;
 import io.servicecomb.common.rest.filter.HttpServerFilter;
 import io.servicecomb.common.rest.locator.OperationLocator;
 import io.servicecomb.common.rest.locator.ServicePathManager;
@@ -32,19 +25,13 @@ import io.servicecomb.core.Const;
 import io.servicecomb.core.CseContext;
 import io.servicecomb.core.Transport;
 import io.servicecomb.core.definition.MicroserviceMeta;
-import io.servicecomb.core.definition.OperationMeta;
 import io.servicecomb.core.invocation.InvocationFactory;
-import io.servicecomb.foundation.metrics.MetricsServoRegistry;
-import io.servicecomb.foundation.metrics.performance.QueueMetrics;
-import io.servicecomb.foundation.metrics.performance.QueueMetricsData;
 import io.servicecomb.foundation.vertx.http.HttpServletRequestEx;
 import io.servicecomb.foundation.vertx.http.HttpServletResponseEx;
 import io.servicecomb.serviceregistry.RegistryUtils;
 import io.servicecomb.swagger.invocation.exception.InvocationException;
 
 public class RestProducerInvocation extends AbstractRestInvocation {
-  private static final Logger LOGGER = LoggerFactory.getLogger(RestProducerInvocation.class);
-
   protected Transport transport;
 
   public void invoke(Transport transport, HttpServletRequestEx requestEx, HttpServletResponseEx responseEx,
@@ -56,7 +43,7 @@ public class RestProducerInvocation extends AbstractRestInvocation {
     requestEx.setAttribute(RestConst.REST_REQUEST, requestEx);
 
     try {
-      this.restOperationMeta = findRestOperation();
+      findRestOperation();
     } catch (InvocationException e) {
       sendFailResponse(e);
       return;
@@ -65,41 +52,7 @@ public class RestProducerInvocation extends AbstractRestInvocation {
     scheduleInvocation();
   }
 
-  protected void scheduleInvocation() {
-    OperationMeta operationMeta = restOperationMeta.getOperationMeta();
-    QueueMetrics metricsData = initMetrics(operationMeta);
-    operationMeta.getExecutor().execute(() -> {
-      synchronized (this.requestEx) {
-        try {
-          if (requestEx.getAttribute(RestConst.REST_REQUEST) != requestEx) {
-            // already timeout
-            // in this time, request maybe recycled and reused by web container, do not use requestEx
-            LOGGER.error("Rest request already timeout, abandon execute, method {}, operation {}.",
-                operationMeta.getHttpMethod(),
-                operationMeta.getMicroserviceQualifiedName());
-            return;
-          }
-
-          runOnExecutor(metricsData);
-        } catch (Throwable e) {
-          LOGGER.error("rest server onRequest error", e);
-          sendFailResponse(e);
-        }
-      }
-    });
-  }
-
-  protected void runOnExecutor(QueueMetrics metricsData) {
-    Object[] args = RestCodec.restToArgs(requestEx, restOperationMeta);
-    this.invocation = InvocationFactory.forProvider(transport.getEndpoint(),
-        restOperationMeta.getOperationMeta(),
-        args);
-    this.invocation.setMetricsData(metricsData);
-    updateMetrics();
-    invoke();
-  }
-
-  protected RestOperationMeta findRestOperation() {
+  protected void findRestOperation() {
     String targetMicroserviceName = requestEx.getHeader(Const.TARGET_MICROSERVICE);
     if (targetMicroserviceName == null) {
       // for compatible
@@ -107,73 +60,18 @@ public class RestProducerInvocation extends AbstractRestInvocation {
     }
     MicroserviceMeta selfMicroserviceMeta =
         CseContext.getInstance().getMicroserviceMetaManager().ensureFindValue(targetMicroserviceName);
-    ServicePathManager servicePathManager = ServicePathManager.getServicePathManager(selfMicroserviceMeta);
-    if (servicePathManager == null) {
-      LOGGER.error("No schema in microservice");
-      throw new InvocationException(Status.NOT_FOUND, Status.NOT_FOUND.getReasonPhrase());
-    }
-
-    OperationLocator locator =
-        servicePathManager.producerLocateOperation(requestEx.getRequestURI(), requestEx.getMethod());
-    requestEx.setAttribute(RestConst.PATH_PARAMETERS, locator.getPathVarMap());
-
-    return locator.getOperation();
+    findRestOperation(selfMicroserviceMeta);
   }
 
   @Override
-  protected void doInvoke() throws Throwable {
-    invocation.next(resp -> {
-      sendResponseQuietly(resp);
-      endMetrics();
-
-    });
+  protected OperationLocator locateOperation(ServicePathManager servicePathManager) {
+    return servicePathManager.producerLocateOperation(requestEx.getRequestURI(), requestEx.getMethod());
   }
 
-  /**
-   * Init the metrics. Note down the queue count and start time.
-   * @param operationMeta Operation data
-   * @return QueueMetrics
-   */
-  private QueueMetrics initMetrics(OperationMeta operationMeta) {
-    QueueMetrics metricsData = new QueueMetrics();
-    metricsData.setQueueStartTime(System.currentTimeMillis());
-    metricsData.setOperQualifiedName(operationMeta.getMicroserviceQualifiedName());
-    QueueMetricsData reqQueue = MetricsServoRegistry.getOrCreateLocalMetrics()
-        .getOrCreateQueueMetrics(operationMeta.getMicroserviceQualifiedName());
-    reqQueue.incrementCountInQueue();
-    return metricsData;
-  }
-
-  /**
-   * Update the queue metrics.
-   */
-  private void updateMetrics() {
-    QueueMetrics metricsData = (QueueMetrics) this.invocation.getMetricsData();
-    if (null != metricsData) {
-      metricsData.setQueueEndTime(System.currentTimeMillis());
-      QueueMetricsData reqQueue = MetricsServoRegistry.getOrCreateLocalMetrics()
-          .getOrCreateQueueMetrics(restOperationMeta.getOperationMeta().getMicroserviceQualifiedName());
-      reqQueue.incrementTotalCount();
-      Long timeInQueue = metricsData.getQueueEndTime() - metricsData.getQueueStartTime();
-      reqQueue.setTotalTime(reqQueue.getTotalTime() + timeInQueue);
-      reqQueue.setMinLifeTimeInQueue(timeInQueue);
-      reqQueue.setMaxLifeTimeInQueue(timeInQueue);
-      reqQueue.decrementCountInQueue();
-    }
-  }
-
-  /**
-   * Prepare the end time of queue metrics.
-   */
-  private void endMetrics() {
-    QueueMetrics metricsData = (QueueMetrics) this.invocation.getMetricsData();
-    if (null != metricsData) {
-      metricsData.setEndOperTime(System.currentTimeMillis());
-      QueueMetricsData reqQueue = MetricsServoRegistry.getOrCreateLocalMetrics()
-          .getOrCreateQueueMetrics(restOperationMeta.getOperationMeta().getMicroserviceQualifiedName());
-      reqQueue.incrementTotalServExecutionCount();
-      reqQueue.setTotalServExecutionTime(
-          reqQueue.getTotalServExecutionTime() + (metricsData.getEndOperTime() - metricsData.getQueueEndTime()));
-    }
+  @Override
+  protected void createInvocation(Object[] args) {
+    this.invocation = InvocationFactory.forProvider(transport.getEndpoint(),
+        restOperationMeta.getOperationMeta(),
+        args);
   }
 }
